# config/swaggerDocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medoc Backend API",
    "version": "1.0.0",
    "description": "API документация для Medoc Backend"
  },
  "servers": [
    {
      "url": "https://medoc.vastness.ru"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Получение файлов для авторизованного пользователя клиники",
        "description": "Возвращает список всех файлов, созданных авторизованным пользователем или публичных файлов.",
        "responses": {
          "200": {
            "description": "Список файлов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "files": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fileName": {
                            "type": "string",
                            "example": "1735277912995-30e99501-af02-4231-a776-1c263608b1ef"
                          },
                          "filePath": {
                            "type": "string",
                            "example": "https://s3.cloud.ru/docuflow-storage/files/1735277912995-30e99501-af02-4231-a776-1c263608b1ef"
                          },
                          "documentTitle": {
                            "type": "string",
                            "example": "Название документа"
                          },
                          "createdBy": {
                            "type": "string",
                            "example": "64d88bce57b0f4e8c7b5d841"
                          },
                          "isPublic": {
                            "type": "boolean",
                            "example": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-12-27T08:40:45.603Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка при получении файлов"
          }
        }
      }
    },
    "/api/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Загрузка файла",
        "description": "Загружает файл в облачное хранилище и сохраняет данные о нём в базе данных.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "documentTitle": {
                    "type": "string",
                    "example": "Название документа"
                  },
                  "isPublic": {
                    "type": "string",
                    "example": "true"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Файл успешно загружен"
          },
          "400": {
            "description": "Ошибка валидации или файл отсутствует в запросе"
          },
          "500": {
            "description": "Ошибка при загрузке файла"
          }
        }
      }
    },
    "/api/files/{fileId}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Удаление файла",
        "description": "Удаляет файл из облачного хранилища и базы данных.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID файла для удаления"
          }
        ],
        "responses": {
          "200": {
            "description": "Файл успешно удалён"
          },
          "404": {
            "description": "Файл не найден"
          },
          "403": {
            "description": "У пользователя нет прав на удаление файла"
          },
          "500": {
            "description": "Ошибка при удалении файла"
          }
        }
      }
    },
...
  }
}